[#1] refactor: 분리했던 slice 파일 삭제, reducer 재통합
