+++ src/redux/reducer/reducer.tsx
@@ -1,6 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 // 코드 흐름을 따라가면서 이해해보시고 추가로 작성해주시면 됩니다.
-
 interface TabState {
   selectedTab: string;
 }
@@ -16,7 +15,6 @@ interface listState {
 interface AccountState {
   pickedAccount: string;
 }
-
 export interface Product {
   // 제품 정보
   id: string; // 제품 ID
@@ -27,7 +25,36 @@ export interface Product {
   thumbnail: string | undefined; // 제품 썸네일 이미지(URL)
   discountRate: number; // 제품 할인율
 }
-
+// 단일 제품 상세 조회 - 예약 정보가 있을 경우 타입
+interface ReservationState {
+  start: string;
+  end: string;
+  isCanceled: boolean;
+  isExpired: boolean;
+}
+// 단일 제품 상세 조회 타입
+interface ProductState {
+  id: string;
+  title: string;
+  price: number;
+  description: string;
+  tags: string[];
+  thumbnail: string | null;
+  photo: string | null;
+  isSoldOut: boolean;
+  reservations: ReservationState[];
+  discountRate: number;
+}
+// 예약 날짜 및 시간
+interface ReserveOptionState {
+  start: string;
+  end: string;
+  timeDiffer: string;
+}
+// 예약 인원
+interface GuestsState {
+  guests: number;
+}
 // 초기값 설정입니다.
 const initialState: TabState = {
   selectedTab: '내 정보',
@@ -45,12 +72,39 @@ const productInitialState: listState = {
     },
   ],
 };
-
+// 예약 정보가 있을 경우 초기값
+const initialReservationState: ReservationState = {
+  start: '',
+  end: '',
+  isCanceled: false,
+  isExpired: false,
+};
+// 단일 제품 상세 조회 초기값
+const initialProductState: ProductState = {
+  id: '',
+  title: '',
+  price: 0,
+  description: '',
+  tags: [],
+  thumbnail: '',
+  photo: '',
+  isSoldOut: false,
+  reservations: [],
+  discountRate: 0,
+};
+const initialReserveOptionState: ReserveOptionState = {
+  start: '',
+  end: '',
+  timeDiffer: '',
+};
+const initialGuestsState: GuestsState = {
+  guests: 1,
+};
 const initialAccountState: AccountState = {
   pickedAccount: '',
 };
 
-const tabSlice = createSlice({
+export const tabSlice = createSlice({
   name: 'tab',
   initialState,
   reducers: {
@@ -62,8 +116,7 @@ const tabSlice = createSlice({
     },
   },
 });
-
-const counterSlice = createSlice({
+export const counterSlice = createSlice({
   name: 'counterSlice',
   initialState: { clickedCounter: 0 },
   reducers: {
@@ -75,8 +128,7 @@ const counterSlice = createSlice({
     },
   },
 });
-
-const selectAccountSlice = createSlice({
+export const selectAccountSlice = createSlice({
   name: 'selectAccountSlice',
   initialState: initialAccountState,
   reducers: {
@@ -85,8 +137,7 @@ const selectAccountSlice = createSlice({
     },
   },
 });
-
-const searchSlice = createSlice({
+export const searchSlice = createSlice({
   name: 'searchSlice',
   initialState: { searchedValue: '' },
   reducers: {
@@ -98,8 +149,7 @@ const searchSlice = createSlice({
     },
   },
 });
-
-const listSlice = createSlice({
+export const listSlice = createSlice({
   name: 'listSlice',
   initialState: productInitialState,
   reducers: {
@@ -108,13 +158,63 @@ const listSlice = createSlice({
     },
   },
 });
-
+export const reservationSlice = createSlice({
+  name: 'reservation',
+  initialState: initialReservationState,
+  reducers: {
+    updateReservation: (state, action: PayloadAction<ReservationState>) => {
+      action.payload;
+    },
+  },
+});
+export const productSlice = createSlice({
+  name: 'product',
+  initialState: initialProductState,
+  reducers: {
+    updateProductDetail: (state, action: PayloadAction<ProductState>) => {
+      state.id = action.payload.id;
+      state.title = action.payload.title;
+      state.price = action.payload.price;
+      state.description = action.payload.description;
+      state.tags = action.payload.tags;
+      state.thumbnail = action.payload.thumbnail;
+      state.photo = action.payload.photo;
+      state.isSoldOut = action.payload.isSoldOut;
+      state.reservations = action.payload.reservations;
+      state.discountRate = action.payload.discountRate;
+    },
+  },
+});
+export const reserveOptionSlice = createSlice({
+  name: 'reserveOption',
+  initialState: initialReserveOptionState,
+  reducers: {
+    selectedDateTime: (state, action: PayloadAction<ReserveOptionState>) => {
+      state.start = action.payload.start;
+      state.end = action.payload.end;
+      state.timeDiffer = action.payload.timeDiffer;
+    },
+  },
+});
+export const guestsSlice = createSlice({
+  name: 'guests',
+  initialState: initialGuestsState,
+  reducers: {
+    selectedGuests: (state, action: PayloadAction<number>) => {
+      state.guests = action.payload;
+    },
+  },
+});
 export interface RootState {
   tabSlice: TabState;
   counterSlice: Counterstate;
   searchSlice: Searchstate;
   selectAccountSlice: AccountState;
   listSlice: listState;
+  productSlice: ProductState;
+  reservationSlice: ReservationState;
+  reserveOptionSlice: ReserveOptionState;
+  guestsSlice: GuestsState;
 }
 
 export const { selectTab } = tabSlice.actions;
@@ -122,6 +222,10 @@ export const { up, down } = counterSlice.actions;
 export const { search, category } = searchSlice.actions;
 export const { selectAccount } = selectAccountSlice.actions;
 export const { plist } = listSlice.actions;
+export const { updateProductDetail } = productSlice.actions;
+export const { updateReservation } = reservationSlice.actions;
+export const { selectedDateTime } = reserveOptionSlice.actions;
+export const { selectedGuests } = guestsSlice.actions;
 
 export default {
   tabSlice: tabSlice.reducer,
@@ -129,4 +233,8 @@ export default {
   searchSlice: searchSlice.reducer,
   selectAccountSlice: selectAccountSlice.reducer,
   listSlice: listSlice.reducer,
+  productSlice: productSlice.reducer,
+  reservationSlice: reservationSlice.reducer,
+  reserveOptionSlice: reserveOptionSlice.reducer,
+  guestsSlice: guestsSlice.reducer,
 };
